refactor: structure of the sign in components
