+++ src/app/(auth)/sign-in/page.tsx
@@ -5,7 +5,7 @@ import SignInModalContainer from "./components/SignInModalContainer";
 
 export default function Component() {
   return (
-    <div className="flex flex-col lg:flex-row items-center justify-center">
+    <div className="flex flex-col xl:flex-row items-center justify-center">
       <BannerContainer />
       <SignInModalContainer />
     </div>
